+++ src/main/java/com/l299l/newbedwars/arena/generators/Generator.java
@@ -1,6 +1,5 @@
 package com.l299l.newbedwars.arena.generators;
 
-import com.l299l.newbedwars.arena.team.Team;
 import org.bukkit.Location;
 
 public interface Generator {