+++ src/main/java/com/l299l/newbedwars/arena/generators/leveling/GeneratorLeveling.java
@@ -2,14 +2,11 @@ package com.l299l.newbedwars.arena.generators.leveling;
 
 import org.bukkit.configuration.file.FileConfiguration;
 
-import java.util.HashMap;
-
 public class GeneratorLeveling {
     private final String id;
     private final GeneratorSettings basicGeneratorSettings;
     private final GeneratorSettings diamondGeneratorSettings;
     private final GeneratorSettings emeraldGeneratorSettings;
-    private final HashMap<String, GeneratorSettings> otherGeneratorSettings;
 
     public GeneratorLeveling(FileConfiguration configuration) {
         if (configuration == null) {
@@ -19,8 +16,6 @@ public class GeneratorLeveling {
         basicGeneratorSettings = new GeneratorSettings(configuration.getConfigurationSection("TeamGenerator"));
         diamondGeneratorSettings = new GeneratorSettings(configuration.getConfigurationSection("DiamondGenerator"));
         emeraldGeneratorSettings = new GeneratorSettings(configuration.getConfigurationSection("EmeraldGenerator"));
-        otherGeneratorSettings = new HashMap<>();
-        boolean otherGeneratorsEnabled = configuration.getBoolean("OtherGenerators.Enabled");
     }
 
     public String getId() {
@@ -37,9 +32,5 @@ public class GeneratorLeveling {
 
     public GeneratorSettings getEmeraldGeneratorSettings() {
         return emeraldGeneratorSettings;
-    }
-
-    public GeneratorSettings getOtherGeneratorSettings(String name) {
-        return otherGeneratorSettings.get(name);
     }
 }