+++ src/main/java/com/l299l/newbedwars/commands/bedwars/admin/setup/normal/SetArenaGuiCommand.java
@@ -29,7 +29,7 @@ public class SetArenaGuiCommand extends SubCommand {
 
     @Override
     public String getExample() {
-        return "/bw setGui shop exampleGui";
+        return "/bw setGui shop exampleShop";
     }
 
     @Override